fix: round lat/lon to 4 decimal places on submission
Browser geolocation returns 13+ decimal places of false precision.
Round client-side to 4 decimals (~11m) before insert, and align
hazards schema to decimal(9,6) to match needs table.

+++ supabase/schema.sql
@@ -134,8 +134,8 @@ CREATE TABLE hazards (
   event_id uuid NOT NULL REFERENCES events(id),
   description text NOT NULL,
   photo_url text,
-  latitude double precision NOT NULL,
-  longitude double precision NOT NULL,
+  latitude decimal(9,6) NOT NULL,
+  longitude decimal(9,6) NOT NULL,
   status hazard_status NOT NULL DEFAULT 'active',
   reported_by text,
   created_at timestamptz NOT NULL DEFAULT now()
